API documentation

Types

CitableText.CtsUrn — Type

Implementation of the CTS URN. See http://cite-architecture.org/ctsurn/.

CitableText.CitableByCtsUrn — Type

A value of CitableBase.CitableTrait that can be used for content citable by CtsUrn.

CitableText.CtsUrnComparable — Type

Value for URN comparison trait on CtsUrnDocStringExtensions

Functions

CitableText.namespace — Function

namespace(u)

Extract CTS namespace from a CtsUrn.

Examples

julia>
namespace(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workcomponent — Function

workcomponent(u)

Extract work component from a CtsUrn.

Examples

julia>
workcomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagecomponent — Function

passagecomponent(u)

Extract passage component from a CtsUrn.

Examples

julia>
passagecomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagedepth — Function

passagedepth(u)

Count number of levels in the passage hierarchy of a CTS URN.

Examples

julia>
passagedepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passageparts — Function

passageparts(u)

Extract an array with all the parts of the passage hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workdepth — Function

workdepth(u)

Count number of levels in the work hierarchy of a CTS URN.

Examples

julia>
workdepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workparts — Function

workparts(u)

Extract an array with all the parts of the work hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.groupid — Function

Extract textgroup identifier from u.

groupid(u)

CitableText.workid — Function

Extract work identifier from u.

workid(u)

CitableBase.versionid — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

versionid(u)

Extract version identifier from u.

versionid(u)

CitableText.exemplarid — Function

Extract exemplar identifier from u.

exemplarid(u)

CitableText.isrange — Function

isrange(psg)

True if passage component string is a range..

Examples

julia>
isrange("1.1-1.10"))

isrange(u)

True if passage component is a range..

Examples

julia>
isrange(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.range_begin — Function

range_begin(s)

Extract first range part from a passage component.

Examples

julia>
range_begin("1.1-1.10"))

range_begin(u)

Extract first range part from a CtsUrn.

Examples

julia>
range_begin(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.range_end — Function

range_end(u)

Extract second range part from a passage component.

Examples

julia>
range_end("1.1-1.10"))

range_end(s)

Extract second range part from a CtsUrn.

Examples

julia>
range_end(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.droppassage — Function

droppassage(u)

Create a new CtsUrn by dropping the passage component of a given CtsUrn.

Examples

julia>
droppassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addpassage — Function

addpassage(u, psg)

Create a new CtsUrn by replacing any passage component in a given CtsUrn with a new passage component.

Examples

julia>
addpassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

Missing docstring.

Missing docstring for dropversion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for addversion. Check Documenter's build log for details.

CitableText.dropexemplar — Function

dropexemplar(u)

Create a new CtsUrn by dropping the exemplar part of a given CtsUrn's work component.

Examples

julia>
dropexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addexemplar — Function

addexemplar(u, ex)

Create a new CtsUrn by replacing any exemplar part of a given CtsUrn's work component with a new exemplar part.

Examples

julia>
addexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.collapsePassageTo — Function

collapsePassageTo(u, count)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy to a given level.

Examples

julia>
collapsePassageTo(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.collapsePassageBy — Function

collapsePassageBy(u, n)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy by a given number of levels.

Examples

julia>
collapsePassageBy(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.workcontains — Function

workcontains(urn1, urn2)

True if work component of urn1 contains or is equal to work component of urn2.

Examples

julia>
workcontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.passagecontains — Function

passagecontains(urn1, urn2)

True if passage component of urn1 contains or is equal to passage component of urn2.

Examples

julia>
passagecontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)