API documentation

API documentation

Types

Functions

User guide

Top-level components

The work component

The passage component

Ranges

Subreferences

CitableText

Types

CitableText.CtsUrn — Type

Implementation of the CTS URN. See http://cite-architecture.org/ctsurn/.

Functions

CitableText.namespace — Function

namespace(u)

Extract CTS namespace from a CtsUrn.

Examples

julia>
namespace(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workcomponent — Function

workcomponent(u)

Extract work component from a CtsUrn.

Examples

julia>
workcomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagecomponent — Function

passagecomponent(u)

Extract passage component from a CtsUrn.

Examples

julia>
passagecomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagedepth — Function

passagedepth(u)

Count number of levels in the passage hierarchy of a CTS URN.

Examples

julia>
passagedepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passageparts — Function

passageparts(u)

Extract an array with all the parts of the passage hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workdepth — Function

workdepth(u)

Count number of levels in the work hierarchy of a CTS URN.

Examples

julia>
workdepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workparts — Function

workparts(u)

Extract an array with all the parts of the work hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.isrange — Function

isrange(psg)

True if passage component string is a range..

Examples

julia>
isrange("1.1-1.10"))

isrange(u)

True if passage component is a range..

Examples

julia>
isrange(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.rangebegin — Function

rangebegin(s)

Extract first range part from a passage component.

Examples

julia>
rangebegin("1.1-1.10"))

rangebegin(u)

Extract first range part from a CtsUrn.

Examples

julia>
rangebegin(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.rangeend — Function

rangeend(u)

Extract second range part from a passage component.

Examples

julia>
rangeend("1.1-1.10"))

rangeend(s)

Extract second range part from a CtsUrn.

Examples

julia>
rangeend(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.hassubref — Function

hassubref(psg)

True if String has a subreference.

Examples

julia>
hassubref(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10@μῆνιν"))

hassubref(u)

True if CtsUrn has a subreference.

Examples

julia>
hassubref(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10@μῆνιν"))

CitableText.subref — Function

subref(s)

Extract subrerence part of a passage string.

Examples

julia>
subref(1.1-1.10@μῆνιν")

CitableText.droppassage — Function

droppassage(u)

Create a new CtsUrn by dropping the passage component of a given CtsUrn.

Examples

julia>
droppassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addpassage — Function

addpassage(u, psg)

Create a new CtsUrn by replacing any passage component in a given CtsUrn with a new passage component.

Examples

julia>
addpassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableBase.dropversion — Function

Catch-all function for dropversion method.

dropversion(u)

dropversion(u)

Create a new CtsUrn by dropping the version part of a given CtsUrn's work component.

Examples

julia>
dropversion(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableBase.addversion — Function

Catch-all function for addversion method.

addversion(u, s)

addversion(u, vers)

Create a new CtsUrn by replacing any version part of a given CtsUrn's work component with a new version part.

Examples

julia>
addversion(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.dropexemplar — Function

dropexemplar(u)

Create a new CtsUrn by dropping the exemplar part of a given CtsUrn's work component.

Examples

julia>
dropexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addexemplar — Function

addexemplar(u, ex)

Create a new CtsUrn by replacing any exemplar part of a given CtsUrn's work component with a new exemplar part.

Examples

julia>
addexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.dropsubref — Function

dropsubref(s)

Remove any subreference substrings in a string. If s is a range expression, separately drop subreference from each of range-begin and range-end parts.

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableText.collapsePassageTo — Function

collapsePassageTo(u, count)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy to a given level.

Examples

julia>
collapsePassageTo(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.collapsePassageBy — Function

collapsePassageBy(u, n)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy by a given number of levels.

Examples

julia>
collapsePassageBy(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.workcontains — Function

workcontains(urn1, urn2)

True if work component of urn1 contains or is equal to work component of urn2.

Examples

julia>
workcontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.passagecontains — Function

passagecontains(urn1, urn2)

True if passage component of urn1 contains or is equal to passage component of urn2.

Examples

julia>
passagecontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.urncontains — Function

urncontains(urn1, urn2)

True if urn1 contains urn2 according to URN logic.

Examples

julia>
urncontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.urnmatches — Function

urnmatches(urn1, urn2)

Override urnmatches function to be true if urn1 and urn1 are urnmatches by URN logic.