API documentation

API documentation

Types

Functions

Index

User guide

CitableText

Types

CitableText.CtsUrn — Type

Implementation of the CTS URN. See http://cite-architecture.org/ctsurn/.

CitableText.CitableNode — Type

A canonically citable text node.

Functions

CitableText.namespace — Method

namespace(u)

Extract CTS namespace from a CtsUrn.

Examples

julia>
namespace(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workcomponent — Method

workcomponent(u)

Extract work component from a CtsUrn.

Examples

julia>
workcomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagecomponent — Method

passagecomponent(u)

Extract passage component from a CtsUrn.

Examples

julia>
passagecomponent(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passagedepth — Method

passagedepth(u)

Count number of levels in the passage hierarchy of a CTS URN.

Examples

julia>
passagedepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.passageparts — Method

passageparts(u)

Extract an array with all the parts of the passage hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workdepth — Method

workdepth(u)

Count number of levels in the work hierarchy of a CTS URN.

Examples

julia>
workdepth(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.workparts — Method

workparts(u)

Extract an array with all the parts of the work hierarchy of a CTS URN.

Examples

julia>
workparts(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1"))

CitableText.isrange — Method

isrange(u)

True if passage component is a range..

Examples

julia>
isrange(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.rangebegin — Method

rangebegin(u)

Extract first range part from a CtsUrn.

Examples

julia>
rangebegin(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.rangeend — Method

rangeend(u)

Extract second range part from a passage component.

Examples

julia>
rangeend("1.1-1.10"))

CitableText.hassubref — Method

hassubref(u)

True if CtsUrn has a subreference.

Examples

julia>
hassubref(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10@μῆνιν"))

CitableText.subref — Method

subref(s)

Extract subrerence part of a passage string.

Examples

julia>
subref(1.1-1.10@μῆνιν")

CitableText.droppassage — Method

droppassage(u)

Create a new CtsUrn by dropping the passage component of a given CtsUrn.

Examples

julia>
droppassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addpassage — Method

addpassage(u, psg)

Create a new CtsUrn by replacing any passage component in a given CtsUrn with a new passage component.

Examples

julia>
addpassage(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.dropversion — Method

dropversion(u)

Create a new CtsUrn by dropping the version part of a given CtsUrn's work component.

Examples

julia>
dropversion(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addversion — Method

addversion(u, vers)

Create a new CtsUrn by replacing any version part of a given CtsUrn's work component with a new version part.

Examples

julia>
addversion(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.dropexemplar — Method

dropexemplar(u)

Create a new CtsUrn by dropping the exemplar part of a given CtsUrn's work component.

Examples

julia>
dropexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.addexemplar — Method

addexemplar(u, ex)

Create a new CtsUrn by replacing any exemplar part of a given CtsUrn's work component with a new exemplar part.

Examples

julia>
addexemplar(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.dropsubref — Method

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableText.collapsePassageTo — Method

collapsePassageTo(u, count)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy to a given level.

Examples

julia>
collapsePassageTo(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.collapsePassageBy — Method

collapsePassageBy(u, n)

Create a new CtsUrn by collapsing a given CtsUrn's passage hierarchy by a given number of levels.

Examples

julia>
collapsePassageBy(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1-1.10"))

CitableText.workcontains — Method

workcontains(urn1, urn2)

True if work component of urn1 contains or is equal to work component of urn2.

Examples

julia>
workcontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.passagecontains — Method

passagecontains(urn1, urn2)

True if passage component of urn1 contains or is equal to passage component of urn2.

Examples

julia>
passagecontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

CitableText.urncontains — Method

urncontains(urn1, urn2)

True if urn1 contains urn2 according to URN logic.

Examples

julia>
urncontains(CtsUrn("urn:cts:greekLit:tlg0012.tlg001.msA:1.1")
CtsUrn("urn:cts:greekLit:tlg0012.tlg001:1")
)

Base.:== — Method

==(urn1, urn2)

Override equality operator for type-specific test.

Base.similar — Method

similar(urn1, urn2)

Override similar function to be true if urn1 and urn1 are similar by URN logic.

CitableText.fromdelimited — Method

fromdelimited(, src)
fromdelimited(, src, delimiter)

Create an instance of type T from a delimited-text String.

CitableText.fromfile — Method

Create an instance of type T from a delimited-text file with header line.

CitableText.fromurl — Method

Create an instance of type T from a URL to a delimited-text file with header line.

CitableText.nextnode — Method

nextnode(n, c)

Retrieve citable node following a given citable node in a corpus. If n is the last node in the corpus, return nothing.

CitableText.prevnode — Method

prevnode(n, c)

Retrieve citable node preceding a given citable node in a corpus. If n is the first node in the corpus, return nothing.

CitableText.catalog — Method

catalog(arr)

Create a CatalogedText from an array of String values.

Missing docstring.

Missing docstring for to_df(arr). Check Documenter's build log for details.

CitableText.combine — Method

combine(c1, c2)

Create a single composite CitableCorpus from two sources.

CitableText.composite_array — Function

composite_array(src_array)
composite_array(src_array, composite)

Create a single composite CitableCorpus from an array of source corpora by recursively adding corpora.

Index

CitableText.CitableNode

CitableText.CtsUrn

Base.:==

Base.similar

CitableText.addexemplar

CitableText.addpassage

CitableText.addversion

CitableText.catalog

CitableText.collapsePassageBy

CitableText.collapsePassageTo

CitableText.combine

CitableText.composite_array

CitableText.dropexemplar

CitableText.droppassage

CitableText.dropsubref

CitableText.dropversion

CitableText.fromdelimited

CitableText.fromfile

CitableText.fromurl

CitableText.hassubref

CitableText.isrange

CitableText.namespace

CitableText.nextnode

CitableText.passagecomponent

CitableText.passagecontains

CitableText.passagedepth

CitableText.passageparts

CitableText.prevnode

CitableText.rangebegin

CitableText.rangeend

CitableText.subref

CitableText.urncontains

CitableText.workcomponent

CitableText.workcontains

CitableText.workdepth

CitableText.workparts